Restructure useGetFullReviewStructure and useGetFullReviewStructureAsync for readability
