+++ src/utils/hooks/useGetFullReviewStructure/index.tsx
@@ -1,0 +1,5 @@
+import useGetFullReviewStructure from './useGetFullReviewStructure'
+import useGetFullReviewStructureAsync from './useGetFullReviewStructureAsync'
+
+export default useGetFullReviewStructure
+export { useGetFullReviewStructureAsync }
